data_server/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
paho-mqtt